Revert "Fix routing and layout issue (#923)" (#930)
This reverts commit 94db2e90da8f7d25f37d2af7e6977cfd7535a454.
